Add utils file
