+++ java/arcs/sdk/wasm/WasmSingletonImpl.kt
@@ -46,6 +46,4 @@ class WasmSingletonImpl<T : WasmEntity>(
             action(entity)
         }
     }
-
-    override fun getContent(): T? = fetch()
 }
